backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
anthropic==0.26.1
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.4
celery==5.4.0
pytest==8.2.1
pytest-asyncio==0.23.6
email-validator==2.1.1
